requirements.txt
imaplib2
openai==0.28
python-dotenv
datasets
langchain
chromadb
sentence-transformers
pydantic
langchain-community
